graphrag_processor/requirements.txt
psycopg2-binary==2.9.9
openai==1.3.0
python-dotenv==1.0.0
networkx==3.1
matplotlib==3.7.2
scipy==1.11.3
pandas==2.0.3
numpy==1.24.3
jsonlines==3.1.0
spacy==3.6.1